.zipcheck/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29
boto3>=1.34
pydantic>=2.5
python-dotenv>=1.0
requests